+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The RSpec integration now inspects hashes and arrays recursively. (Like rspec does by default)
 - RSpec integration no longer breaks description output of matchers when using multi matchers (like .all or .and)
 - The hash differ now deals with recursive hashes and arrays
+- RSpec integration no longer breaks description output of matchers that are part of a diff inside an array or hash. (like when doing `match([have_attributes(...)])`)
 
 ## [0.3.0-rc1] - 2024-04-02