+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0.rc2] - 2024-04-05
+
 ### Changed
 
 - Rework how hashdiff's output gets printed.